backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
cryptography==41.0.7
python-dotenv==1.0.0
sqlite3-simple==0.1.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
datetime
json5==0.9.14